+++ packages/adapter-mongoose/lib/adapter-mongoose.js
@@ -8,19 +8,11 @@ const {
   mapKeyNames,
   identity,
   mergeWhereClause,
-} = require('@keystone-alpha/utils');
+} = require('@keystonejs/utils');
 
-const {
-  BaseKeystoneAdapter,
-  BaseListAdapter,
-  BaseFieldAdapter,
-} = require('@keystone-alpha/keystone');
-const {
-  queryParser,
-  pipelineBuilder,
-  mutationBuilder,
-} = require('@keystone-alpha/mongo-join-builder');
-const logger = require('@keystone-alpha/logger').logger('mongoose');
+const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('@keystonejs/keystone');
+const { queryParser, pipelineBuilder, mutationBuilder } = require('@keystonejs/mongo-join-builder');
+const logger = require('@keystonejs/logger').logger('mongoose');
 
 const slugify = require('@sindresorhus/slugify');
 
@@ -82,7 +74,7 @@ class MongooseAdapter extends BaseKeystoneAdapter {
 
   getDefaultPrimaryKeyConfig() {
     // Required here due to circular refs
-    const { MongoId } = require('@keystone-alpha/fields-mongoid');
+    const { MongoId } = require('@keystonejs/fields-mongoid');
     return MongoId.primaryKeyDefaults[this.name].getConfig();
   }
 }
@@ -211,7 +203,7 @@ class MongooseListAdapter extends BaseListAdapter {
         { include: from.fromField }
       );
       if (ids.length) {
-        args = mergeWhereClause(args, { id: { $in: ids[0][from.fromField] } });
+        args = mergeWhereClause(args, { id: { $in: ids[0][from.fromField] || [] } });
       }
     }
     function graphQlQueryToMongoJoinQuery(query) {