+++ packages/app-admin-ui/client/pages/List/url-state.js
@@ -5,7 +5,7 @@ import * as React from 'react';
 
 import List from '../../classes/List';
 // $FlowFixMe flow doesn't recognise the `*/Controller` entry point
-import type { FieldControllerType } from '@keystone-alpha/fields/Controller';
+import type { FieldControllerType } from '@keystonejs/fields/Controller';
 import { pseudoLabelField } from './FieldSelect';
 import type { AdminMeta } from '../../providers/AdminMeta';