+++ packages/fields-auto-increment/src/Implementation.js
@@ -1,5 +1,5 @@
-import { Implementation } from '@keystone-alpha/fields';
-import { KnexFieldAdapter } from '@keystone-alpha/adapter-knex';
+import { Implementation } from '@keystonejs/fields';
+import { KnexFieldAdapter } from '@keystonejs/adapter-knex';
 
 export class AutoIncrementImplementation extends Implementation {
   constructor(path, config = {}, context = {}) {