+++ test-projects/basic/CHANGELOG.md
@@ -1,4 +1,55 @@
+# @keystonejs/cypress-project-basic
+
+## 5.0.1
+
+### Patch Changes
+
+- [`ce6e98f2`](https://github.com/keystonejs/keystone/commit/ce6e98f2cf63f5c27a7433e932e8bfbd00376874) [#1836](https://github.com/keystonejs/keystone/pull/1836) Thanks [@timleslie](https://github.com/timleslie)! - Deployed cypress-file-upload instead of a custom implementation
+
+- Updated dependencies [[`209b7078`](https://github.com/keystonejs/keystone/commit/209b7078c7fa4f4d87568c58cb6cb6ad8162fe46), [`19b08a30`](https://github.com/keystonejs/keystone/commit/19b08a30b3dbfb7c7a0056f210769bbf6e171c85), [`3c19cddd`](https://github.com/keystonejs/keystone/commit/3c19cddd0b8b8d1e17385a01a813a9e84ec14bb5)]:
+  - @keystonejs/fields@5.0.1
+  - @keystonejs/file-adapters@5.0.1
+  - @keystonejs/app-admin-ui@5.0.1
+  - @keystonejs/adapter-mongoose@5.0.1
+
+## 5.0.0
+
+### Major Changes
+
+- [`7b4ed362`](https://github.com/keystonejs/keystone-5/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310) [#1821](https://github.com/keystonejs/keystone-5/pull/1821) Thanks [@jesstelford](https://github.com/jesstelford)! - Release @keystonejs/\* packages (つ＾ ◡ ＾)つ
+
+  - This is the first release of `@keystonejs/*` packages (previously `@keystone-alpha/*`).
+  - All packages in the `@keystone-alpha` namespace are now available in the `@keystonejs` namespace, starting at version `5.0.0`.
+  - To upgrade your project you must update any `@keystone-alpha/*` dependencies in `package.json` to point to `"@keystonejs/*": "^5.0.0"` and update any `require`/`import` statements in your code.
+
+### Patch Changes
+
+- Updated dependencies [[`7b4ed362`](https://github.com/keystonejs/keystone-5/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310)]:
+  - @keystonejs/adapter-mongoose@5.0.0
+  - @keystonejs/app-admin-ui@5.0.0
+  - @keystonejs/app-graphql@5.0.0
+  - @keystonejs/app-static@5.0.0
+  - @keystonejs/field-content@5.0.0
+  - @keystonejs/fields-markdown@5.0.0
+  - @keystonejs/fields@5.0.0
+  - @keystonejs/file-adapters@5.0.0
+  - @keystonejs/keystone@5.0.0
+  - @keystonejs/oembed-adapters@5.0.0
+
 # @keystone-alpha/cypress-project-basic
+
+## 1.3.18
+
+### Patch Changes
+
+- Updated dependencies [[`0a36b0f4`](https://github.com/keystonejs/keystone-5/commit/0a36b0f403da73a76106b5e14940a789466b4f94), [`7129c887`](https://github.com/keystonejs/keystone-5/commit/7129c8878a825d961f2772be497dcd5bd6b2b697), [`3bc02545`](https://github.com/keystonejs/keystone-5/commit/3bc025452fb8e6e69790bdbee032ddfdeeb7dabb), [`768420f5`](https://github.com/keystonejs/keystone-5/commit/768420f567c244d57a4e2a3aaafe628ea9813d9d), [`a48281ba`](https://github.com/keystonejs/keystone-5/commit/a48281ba605bf5bebc89fcbb36d3e69c17182eec), [`a8ee0179`](https://github.com/keystonejs/keystone-5/commit/a8ee0179842f790dd3b5d4aae3524793e752ee26), [`effc1f63`](https://github.com/keystonejs/keystone-5/commit/effc1f639d5824720b7a9d82c2ee881d77acb901)]:
+  - @keystone-alpha/keystone@16.1.0
+  - @keystone-alpha/app-graphql@8.2.1
+  - @keystone-alpha/adapter-mongoose@6.0.1
+  - @keystone-alpha/app-admin-ui@5.10.3
+  - @keystone-alpha/fields@15.0.0
+  - @keystone-alpha/field-content@3.1.4
+  - @keystone-alpha/fields-markdown@1.0.8
 
 ## 1.3.17